+++ insts/jg/ELL/ELL.cpp
@@ -84,7 +84,7 @@ ELL::~ELL()
 }
 
 
-int ELL::init(float p[], int n_args)
+int ELL::init(double p[], int n_args)
 {
    int   n;
    float outskip, inskip, dur, ringdur;
